64 bit: make tango compilable on 64 bit systems. Most unittests pass, those that don't are very likely dmd bugs

+++ tango/io/device/Device.d
@@ -311,7 +311,7 @@ class Device : Conduit, ISelectable
 
                 override size_t write (const(void)[] src)
                 {
-                        int written = posix.write (handle, src.ptr, src.length);
+                        size_t written = posix.write (handle, src.ptr, src.length);
                         if (written is -1)
                             error;
                         return written;
